moved parser into new structure. making core library a java library while the jsonparser is now an android library and going to add fastjson as another module
