+++ Parser-Compiler/gradle.properties
@@ -1,3 +1,3 @@
 POM_NAME=Parser-Compiler
-POM_PACKAGING=aar
+POM_PACKAGING=jar
 POM_ARTIFACT_ID=Parser-Compiler
