+++ Parser-Core/gradle.properties
@@ -1,3 +1,3 @@
 POM_NAME=Parser-Core
-POM_PACKAGING=aar
+POM_PACKAGING=jar
 POM_ARTIFACT_ID=Parser-Core
