+++ Parser/src/main/java/com/raizlabs/android/parser/ObjectParser.java
@@ -7,9 +7,7 @@ package com.raizlabs.android.parser;
 public interface ObjectParser<ParseableClass> {
 
     /**
-     * Returns a new instance of the {@link ParseableClass} without using reflection.
-     *
-     * @return
+     * @return a new instance of the {@link ParseableClass} without using reflection.
      */
     public ParseableClass getInstance();
 
@@ -17,7 +15,7 @@ public interface ObjectParser<ParseableClass> {
      * Will set all of the fields for the {@link ParseableClass} that define a {@link com.raizlabs.android.parser.core.Key} attribute.
      *
      * @param parseable     The instance that we are parsing
-     * @param dataInstance The singular data type that we're parsing. E.g. {@link org.json.JSONObject}
+     * @param dataInstance The singular data type that we're parsing. E.g. JSONObject
      * @param parser       The parser we're using to parse the dataInstance into the parsible
      */
     public void parse(ParseableClass parseable, Object dataInstance, Parser parser);
