+++ parser-jsonparser/build.gradle
@@ -1,0 +1,30 @@
+apply plugin: 'com.android.library'
+apply plugin: 'android-apt'
+
+
+android {
+    compileSdkVersion 21
+    buildToolsVersion "21.1.1"
+
+    defaultConfig {
+        minSdkVersion 3
+        targetSdkVersion 21
+    }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_7
+        targetCompatibility JavaVersion.VERSION_1_7
+    }
+
+    packagingOptions {
+        exclude 'META-INF/services/javax.annotation.processing.Processor'
+    }
+}
+
+dependencies {
+    compile project(parser_project_prefix + "Parser")
+    androidTestApt project(parser_project_prefix + "Parser-Compiler")
+}
+
+apply from: '../maven_push.gradle'
+apply from: '../android-artifacts.gradle'
