+++ parser-jsonparser/gradle.properties
@@ -1,0 +1,3 @@
+POM_NAME=Parser-JsonParser
+POM_PACKAGING=aar
+POM_ARTIFACT_ID=Parser-JsonParser
