+++ parser-jsonparser/src/androidTest/AndroidManifest.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.raizlabs.android.parser.test" >
+          package="com.raizlabs.android.parser.jsonparser" >
 
     <instrumentation
             android:name="android.test.InstrumentationTestRunner"
-            android:targetPackage="com.raizlabs.android.parser.test" />
+            android:targetPackage="com.raizlabs.android.parser.jsonparser.test" />
 
     <application />
