+++ parser-jsonparser/src/main/AndroidManifest.xml
@@ -1,0 +1,6 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+          package="com.raizlabs.android.parser.jsonparser">
+
+    <application/>
+
+</manifest>
