+++ settings.gradle
@@ -1,2 +1,2 @@
-include ':app' ,':Parser', ':Parser-Compiler', ':Parser-Core'
+include ':app', ':Parser', ':Parser-Compiler', ':Parser-Core', 'Parser-JsonParser'
